+++ proofs/manager.php
@@ -23,7 +23,7 @@ use Properties\Formal\ORM\{
 };
 use Formal\AccessLayer\{
     Query\DropTable,
-    Query\SQL,
+    Query\Delete,
     Table,
 };
 use Innmind\OperatingSystem\Factory;
@@ -44,7 +44,7 @@ return static function() {
 
             $assert->same($repository1, $repository2);
         },
-    );
+    )->tag(Storage::filesystem);
     yield test(
         'Manager::repository() returns an instance per class',
         static function($assert) {
@@ -57,7 +57,7 @@ return static function() {
                 ->not()
                 ->same($repositoryB);
         },
-    );
+    )->tag(Storage::filesystem);
 
     yield test(
         'Nested transactions are forbidden',
@@ -74,7 +74,7 @@ return static function() {
                 'Nested transactions not allowed',
             );
         },
-    );
+    )->tag(Storage::filesystem);
 
     yield properties(
         'Filesystem properties',
@@ -86,7 +86,7 @@ return static function() {
                 SortableType::of(...),
             )),
         )),
-    );
+    )->tag(Storage::filesystem);
 
     foreach (Properties::alwaysApplicable() as $property) {
         yield property(
@@ -98,46 +98,61 @@ return static function() {
                     SortableType::of(...),
                 )),
             )),
-        )->named('Filesystem');
+        )
+            ->named('Filesystem')
+            ->tag(Storage::filesystem);
     }
 
     $os = Factory::build();
-
-    $port = \getenv('DB_PORT') ?: '3306';
-    $connection = $os->remote()->sql(Url::of("mysql://root:root@127.0.0.1:$port/example"));
     $aggregates = Aggregates::of(Types::of(
-        Type\PointInTimeType::of(new Clock),
+        Type\PointInTimeType::of($os->clock()),
         SortableType::of(...),
     ));
-    $connection(DropTable::ifExists(Table\Name::of('user_roles')));
-    $connection(DropTable::ifExists(Table\Name::of('user_addresses')));
-    $connection(DropTable::ifExists(Table\Name::of('user_mainAddress')));
-    $connection(DropTable::ifExists(Table\Name::of('user_billingAddress')));
-    $connection(DropTable::ifExists(Table\Name::of('user')));
-    $_ = Adapter\SQL\ShowCreateTable::of($aggregates)(User::class)->foreach($connection);
 
-    $setup = static function() use ($connection, $aggregates) {
-        $connection(SQL::of('DELETE FROM user_roles'));
-        $connection(SQL::of('DELETE FROM user_addresses'));
-        $connection(SQL::of('DELETE FROM user_mainAddress'));
-        $connection(SQL::of('DELETE FROM user_billingAddress'));
-        $connection(SQL::of('DELETE FROM user'));
+    $sql = static function(Url $dsn, string $driver) use ($os, $aggregates) {
+        $connection = $os->remote()->sql($dsn);
+        $connection(DropTable::ifExists(Table\Name::of('user_roles')));
+        $connection(DropTable::ifExists(Table\Name::of('user_addresses')));
+        $connection(DropTable::ifExists(Table\Name::of('user_mainAddress')));
+        $connection(DropTable::ifExists(Table\Name::of('user_billingAddress')));
+        $connection(DropTable::ifExists(Table\Name::of('user')));
+        $_ = Adapter\SQL\ShowCreateTable::of($aggregates)(User::class)->foreach($connection);
 
-        return Manager::sql($connection, $aggregates);
+        $setup = static function() use ($connection, $aggregates) {
+            $connection(Delete::from(Table\Name::of('user_roles')));
+            $connection(Delete::from(Table\Name::of('user_addresses')));
+            $connection(Delete::from(Table\Name::of('user_mainAddress')));
+            $connection(Delete::from(Table\Name::of('user_billingAddress')));
+            $connection(Delete::from(Table\Name::of('user')));
+
+            return Manager::sql($connection, $aggregates);
+        };
+
+        yield properties(
+            "SQL properties($driver)",
+            Properties::any(),
+            Set\Call::of($setup),
+        )->tag(Storage::sql);
+
+        foreach (Properties::alwaysApplicable() as $property) {
+            yield property(
+                $property,
+                Set\Call::of($setup),
+            )
+                ->named("SQL($driver)")
+                ->tag(Storage::sql);
+        }
     };
 
-    yield properties(
-        'SQL properties',
-        Properties::any(),
-        Set\Call::of($setup),
-    );
+    $port = \getenv('DB_PORT') ?: '3306';
+    $dsn = Url::of("mysql://root:root@127.0.0.1:$port/example");
 
-    foreach (Properties::alwaysApplicable() as $property) {
-        yield property(
-            $property,
-            Set\Call::of($setup),
-        )->named('SQL');
-    }
+    yield from $sql($dsn, 'mysql');
+
+    $port = \getenv('POSTGRES_DB_PORT') ?: '5432';
+    $dsn = Url::of("pgsql://root:root@127.0.0.1:$port/example");
+
+    yield from $sql($dsn, 'postgres');
 
     $port = \getenv('ES_PORT') ?: '9200';
     $url = Url::of("http://127.0.0.1:$port/");
@@ -172,7 +187,7 @@ return static function() {
         'Elasticsearch properties',
         Properties::any(Properties::withoutTransactions()),
         Set\Call::of($setup),
-    );
+    )->tag(Storage::elasticsearch);
 
     foreach (Properties::alwaysApplicable() as $property) {
         if (\in_array(
@@ -186,6 +201,8 @@ return static function() {
         yield property(
             $property,
             Set\Call::of($setup),
-        )->named('Elasticsearch');
+        )
+            ->named('Elasticsearch')
+            ->tag(Storage::elasticsearch);
     }
 };